clean the implementation of Dictionary.
