+++ pixels-core/src/main/java/io/pixelsdb/pixels/core/utils/StringRedBlackTree.java
@@ -26,10 +26,11 @@ import java.nio.charset.CharacterCodingException;
 import java.util.Arrays;
 
 /**
+ * This id derived from the StringRedBlockTree implementation in Apache Orc.
  * A red-black tree that stores strings. The strings are stored as UTF-8 bytes
  * and an offset for each entry.
  */
-public class StringRedBlackTree extends RedBlackTree
+public class StringRedBlackTree extends RedBlackTree implements Dictionary
 {
     private final DynamicByteArray byteArray = new DynamicByteArray();
     private final DynamicIntArray keyOffsets;
@@ -42,6 +43,7 @@ public class StringRedBlackTree extends RedBlackTree
         keyOffsets = new DynamicIntArray(initialCapacity);
     }
 
+    @Override
     public int add(String value)
     {
         try
@@ -62,6 +64,13 @@ public class StringRedBlackTree extends RedBlackTree
         }
     }
 
+    @Override
+    public int add(byte[] keyContent)
+    {
+        return add(keyContent, 0, keyContent.length);
+    }
+
+    @Override
     public int add(byte[] value, int offset, int length)
     {
         setCapacity(length, false);
@@ -89,23 +98,6 @@ public class StringRedBlackTree extends RedBlackTree
         }
     }
 
-//    private ByteBuffer encode(String string, boolean replace)
-//            throws CharacterCodingException
-//    {
-//        CharsetEncoder encoder = ENCODER_FACTORY.get();
-//        if (replace) {
-//            encoder.onMalformedInput(CodingErrorAction.REPLACE);
-//            encoder.onUnmappableCharacter(CodingErrorAction.REPLACE);
-//        }
-//        ByteBuffer bytes =
-//                encoder.encode(CharBuffer.wrap(string.toCharArray()));
-//        if (replace) {
-//            encoder.onMalformedInput(CodingErrorAction.REPORT);
-//            encoder.onUnmappableCharacter(CodingErrorAction.REPORT);
-//        }
-//        return bytes;
-//    }
-
     // compare newKey with position specified key
     @Override
     protected int compareValue(int position)
@@ -124,120 +116,23 @@ public class StringRedBlackTree extends RedBlackTree
                 start, end - start);
     }
 
-    /**
-     * The information about each node.
-     */
-    public interface VisitorContext
-    {
-        /**
-         * Get the position where the key was originally added.
-         *
-         * @return the number returned by add.
-         */
-        int getOriginalPosition();
-
-        /**
-         * Write the bytes for the string to the given output stream.
-         *
-         * @param out the stream to write to.
-         * @throws IOException
-         */
-        void writeBytes(OutputStream out)
-                throws IOException;
-
-        /**
-         * Get the original string.
-         *
-         * @return the string
-         */
-        String getString();
-
-        /**
-         * Get the number of bytes.
-         *
-         * @return the string's length in bytes
-         */
-        int getLength();
-    }
-
-    /**
-     * The interface for visitors.
-     */
-    public interface Visitor
-    {
-        /**
-         * Called once for each node of the tree in sort order.
-         *
-         * @param context the information about each node
-         * @throws IOException
-         */
-        void visit(VisitorContext context)
-                throws IOException;
-    }
-
-    private class VisitorContextImpl implements VisitorContext
-    {
-        private int originalPosition;
-        private int start;
-        private int end;
-        private final Text text = new Text();
-
-        public int getOriginalPosition()
-        {
-            return originalPosition;
-        }
-
-        public String getString()
-        {
-            byteArray.setText(text, start, end - start);
-            return text.toString();
-        }
-
-        public void writeBytes(OutputStream out)
-                throws IOException
-        {
-            byteArray.write(out, start, end - start);
-        }
-
-        public int getLength()
-        {
-            return end - start;
-        }
-
-        void setPosition(int position)
-        {
-            originalPosition = position;
-            start = keyOffsets.get(originalPosition);
-            if (position + 1 == keyOffsets.size())
-            {
-                end = byteArray.size();
-            }
-            else
-            {
-                end = keyOffsets.get(originalPosition + 1);
-            }
-        }
-    }
-
-    private void recurse(int node, Visitor visitor, VisitorContextImpl context
-    )
-            throws IOException
+    private void recurse(int node, Visitor visitor, VisitorContextImpl context) throws IOException
     {
         if (node != NULL)
         {
             recurse(getLeft(node), visitor, context);
-            context.setPosition(node);
+            context.setKeyPosition(node);
             visitor.visit(context);
             recurse(getRight(node), visitor, context);
         }
     }
 
     /**
-     * Visit all of the nodes in the tree in sorted order.
-     *
+     * Visit all the nodes in the tree in sorted order.
      * @param visitor the action to be applied to each node
      * @throws IOException
      */
+    @Override
     public void visit(Visitor visitor)
             throws IOException
     {
@@ -245,8 +140,9 @@ public class StringRedBlackTree extends RedBlackTree
     }
 
     /**
-     * Reset the table to empty.
+     * Reset the dictionary to empty.
      */
+    @Override
     public void clear()
     {
         super.clear();
@@ -254,30 +150,43 @@ public class StringRedBlackTree extends RedBlackTree
         keyOffsets.clear();
     }
 
-    public String getString(int originalPosition)
+    private class VisitorContextImpl implements VisitorContext
     {
-        Text result = new Text();
-        int offset = keyOffsets.get(originalPosition);
-        int length;
-        if (originalPosition + 1 == keyOffsets.size())
-        {
-            length = byteArray.size() - offset;
-        }
-        else
-        {
-            length = keyOffsets.get(originalPosition + 1) - offset;
-        }
-        byteArray.setText(result, offset, length);
-        return result.toString();
-    }
+        private int keyPosition;
+        private int start;
+        private int end;
 
-    /**
-     * Get the size of the character data in the table.
-     *
-     * @return the bytes used by the table
-     */
-    public int getCharacterSize()
-    {
-        return byteArray.size();
+        @Override
+        public int getKeyPosition()
+        {
+            return keyPosition;
+        }
+
+        @Override
+        public void writeBytes(OutputStream out)
+                throws IOException
+        {
+            byteArray.write(out, start, end - start);
+        }
+
+        @Override
+        public int getLength()
+        {
+            return end - start;
+        }
+
+        void setKeyPosition(int keyPosition)
+        {
+            this.keyPosition = keyPosition;
+            start = keyOffsets.get(keyPosition);
+            if (keyPosition + 1 == keyOffsets.size())
+            {
+                end = byteArray.size();
+            }
+            else
+            {
+                end = keyOffsets.get(keyPosition + 1);
+            }
+        }
     }
 }
